config: add newlines after the k8s kinds
This fixes an issue where the default config map will incorrectly match certain
objects such as ClusterRole when only ClusterRoleBinding is present.  The reason
for this is because:

    "apiVersion: rbac.authorization.k8s.io/v1beta1\nkind: ClusterRole"

is a substring of:

    "apiVersion: rbac.authorization.k8s.io/v1beta1\nkind: ClusterRoleBinding"

This means that if you have a helm config with only ClusterRoleBinding, the console
output will print a message that `ClusterRole` was matched and fixed.
Adding a newline to the end of each match string ensures that superstrings will not
be matched.

Signed-off-by: Paul Gier <paul.gier@datastax.com>